Refactor quote pricing helpers and remove test-only exports
