+++ src/_lib/public/utils/quote-price-utils.js
@@ -1,8 +1,8 @@
 // Quote price display utilities
 // Renders a price summary for cart items with hire pricing
 
-import { isHireItem } from "#public/cart/hire-calculator.js";
 import { getRadioValue } from "#public/cart/quote-steps.js";
+import { getPriceForDays, sanitizeItemName } from "#public/utils/quote-pricing.js";
 import { formatPrice, getCart } from "#public/utils/cart-utils.js";
 import { IDS } from "#public/utils/selectors.js";
 import { getTemplate } from "#public/utils/template.js";
@@ -15,34 +15,12 @@ import {
   uniqueBy,
 } from "#toolkit/fp/array.js";
 
-const parsePrice = (priceStr) => {
-  if (typeof priceStr === "number") return priceStr;
-  if (!priceStr) return 0;
-  const matches = String(priceStr).match(/[\d.]+/);
-  return matches ? Number.parseFloat(matches[0]) : 0;
-};
-
 const sum = reduce((acc, n) => acc + n, 0);
-const HIRE_DAY_SUFFIX = /\s-\s\d+\sday(?:s)?$/i;
-
-// Returns null if hire item lacks price for that day count
-const getPriceForDays = (days) => (item) => {
-  if (!isHireItem(item)) {
-    return item.unit_price * item.quantity;
-  }
-  const price = item.hire_prices[days];
-  return price ? parsePrice(price) * item.quantity : null;
-};
 
 const formatItemName = (item) =>
   item.quantity > 1
     ? `${sanitizeItemName(item)} (×${item.quantity})`
     : sanitizeItemName(item);
-
-const sanitizeItemName = (item) =>
-  isHireItem(item)
-    ? item.item_name.replace(HIRE_DAY_SUFFIX, "")
-    : item.item_name;
 
 const formatItemPrice = (price) =>
   price === null ? "TBC" : formatPrice(price);
@@ -194,4 +172,3 @@ const setupDetailsBlurHandlers = (getDays = () => 1) => {
 };
 
 export { setupDetailsBlurHandlers, updateQuotePrice };
-export { getPriceForDays, sanitizeItemName };
